ai-service/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Machine Learning
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3

# NLP
nltk==3.8.1
Sastrawi==1.0.1

# Optional: Deep Learning (uncomment if using IndoBERT)
# transformers==4.35.2
# torch==2.1.1

# Utilities
python-dotenv==1.0.0
joblib==1.3.2

# LLM Integration
openai==1.54.0

# CORS
python-jose[cryptography]==3.3.0
 
# Cache
redis==5.0.1
